toolkit/components/passwordmgr/test/test_xml_load.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Login Manager</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>  
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <script type="text/javascript" src="prompt_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Login Manager test: XML prompt
<p id="display"></p>

<div id="content" style="display: none">
  <iframe id="iframe"></iframe>
</div>

<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Login Manager: XML prompts. **/
var pwmgr, login1, login2;

function initLogins() {
  pwmgr = SpecialPowers.Cc["@mozilla.org/login-manager;1"]
                       .getService(Ci.nsILoginManager);

  login1 = SpecialPowers.Cc["@mozilla.org/login-manager/loginInfo;1"]
                        .createInstance(Ci.nsILoginInfo);
  login2 = SpecialPowers.Cc["@mozilla.org/login-manager/loginInfo;1"]
                        .createInstance(Ci.nsILoginInfo);

  login1.init("http://mochi.test:8888", null, "xml",
               "xmluser1", "xmlpass1", "", "");
  login2.init("http://mochi.test:8888", null, "xml2",
               "xmluser2", "xmlpass2", "", "");

  pwmgr.addLogin(login1);
  pwmgr.addLogin(login2);
}

function finishTest() {
  ok(true, "finishTest removing testing logins...");
  pwmgr.removeLogin(login1);
  pwmgr.removeLogin(login2);

  SimpleTest.finish();
}

function handleDialog(doc, testNum) {
  ok(true, "handleDialog running for test " + testNum);

  var clickOK = true;
  var userfield = doc.getElementById("loginTextbox");
  var passfield = doc.getElementById("password1Textbox");
  var username = userfield.getAttribute("value");
  var password = passfield.getAttribute("value");
  var dialog    = doc.getElementById("commonDialog");

  switch(testNum) {
    case 1:
        is(username, "xmluser1", "Checking provided username");
        is(password, "xmlpass1", "Checking provided password");
        break;

    case 2:
        is(username, "xmluser2", "Checking provided username");
        is(password, "xmlpass2", "Checking provided password");

        // Check that the dialog has the correct parent
        ok(doc.defaultView.opener, "dialog has opener");
        // Not using is() because its "expected" text doesn't deal
        // with window objects very well 
        // Disabled due to Bug 718543.
        // ok(doc.defaultView.opener == window, "dialog's opener is correct");

        break;

    default:
        ok(false, "Uhh, unhandled switch for testNum #" + testNum);
        break;
  }

  // Explicitly cancel the dialog and report a fail in this failure
  // case, rather than letting the dialog get stuck due to an auth
  // failure and having the test timeout.
  if (!username && !password) {
      ok(false, "No values prefilled");
      clickOK = false;
  }

  if (clickOK)
    dialog.acceptDialog();
  else
    dialog.cancelDialog();

  ok(true, "handleDialog done");
  didDialog = true;
}

var newWin;
function xmlLoad(responseDoc) {
  ok(true, "xmlLoad running for test " + testNum);

  // The server echos back the user/pass it received.
  var username = responseDoc.getElementById("user").textContent;
  var password = responseDoc.getElementById("pass").textContent;
  var authok = responseDoc.getElementById("ok").textContent;

  switch(testNum) {
    case 1:
        is(username, "xmluser1", "Checking provided username");
        is(password, "xmlpass1", "Checking provided password");
        break;

    case 2:
        is(username, "xmluser2", "Checking provided username");
        is(password, "xmlpass2", "Checking provided password");

        newWin.close();
        break;

    default:
        ok(false, "Uhh, unhandled switch for testNum #" + testNum);
        break;
  }
  
  doTest();
}

function doTest() {
  switch(++testNum) {
    case 1:
        startCallbackTimer();
        makeRequest("authenticate.sjs?user=xmluser1&pass=xmlpass1&realm=xml");
        break;

    case 2:
        // Test correct parenting, by opening another window and
        // making sure the prompt's opener is correct
        newWin = window.open();
        newWin.focus();
        startCallbackTimer();
        makeRequest("authenticate.sjs?user=xmluser2&pass=xmlpass2&realm=xml2");
        break;

    default:
        finishTest();
  }
}

function makeRequest(uri) {
  var xmlDoc = document.implementation.createDocument("", "test", null);

  function documentLoaded(e) {
      xmlLoad(xmlDoc);
  }
  xmlDoc.addEventListener("load", documentLoaded, false);
  xmlDoc.load(uri);
}


initLogins();

// clear plain HTTP auth sessions before the test, to allow
// running them more than once.
var authMgr = SpecialPowers.Cc['@mozilla.org/network/http-auth-manager;1']
                           .getService(SpecialPowers.Ci.nsIHttpAuthManager);
authMgr.clearAll();

// start the tests
testNum = 0;
doTest();

SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
